requirements.txt
# KIS API
requests==2.31.0
websockets==12.0

# AI Agents
anthropic>=0.96.0
openai>=2.32.0
google-genai>=1.73.0

# Data & Indicators
pandas==2.2.2
numpy==1.26.4
ta==0.11.0          # Technical Analysis library

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
aiohttp==3.10.5
asyncio-throttle==1.0.2
